apps/mimic/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
httpx==0.25.1
cryptography==41.0.7
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0
celery==5.3.4
redis==5.0.1
jinja2==3.1.2
stripe==14.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.1
